client/requirements.txt
impacket
lxml
pefile
pynacl
pywin32 ; sys_platform == "win32"